chore: removed logs and added hooks
